refactor project - trying to get user login to work
