+++ package.json
@@ -27,6 +27,7 @@
     "react-transition-group": "^2.2.1",
     "redux": "^3.7.2",
     "redux-promise": "^0.5.3",
+    "redux-thunk": "^2.2.0",
     "semantic-ui-react": "^0.77.0"
   },
   "scripts": {
