+++ src/App.css
@@ -242,7 +242,7 @@ Language Components
 
 
 /************************************************
-LANDING PAGE
+LANDING PAGE AKA FIXED MENU
 ************************************************/
 
 .LoginButton {
