+++ src/actions/index.js
@@ -3,7 +3,7 @@ import axios from 'axios';
 export const FETCH_FLASHCARDS = 'fetch_flashcards';
 export const FETCH_USER = 'fetch_user';
 
-const ROOT_URL = 'http://localhost:8080';
+const ROOT_URL = 'http://localhost:3001';
 
 export const fetchCards = () => {
   const request = axios.get(`${ROOT_URL}/cards`)
@@ -15,9 +15,9 @@ export const fetchCards = () => {
 }
 
 export const fetchUser = () => async dispatch => {
-  const res = await axios.get('/api/current_user');
+  const res = await axios.get(`/api/current_user`);
 
-  dispatch({ type: FETCH_USER, payload: res.data});
+  dispatch({ type: FETCH_USER, payload: res.data });
 };
